requirements.txt
Django==1.9.7
djangorestframework==3.4.0
Pillow==3.3.0
redis==2.10.5
django-redis-sessions==0.5.6
django-phonenumber-field==1.1.0
